robot_body/src/modules/vision/test/requirements.txt
# Web 服务相关
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# 图像处理相关
opencv-python>=4.8.1.78
numpy>=1.24.3
pillow>=10.0.0

# 深度学习相关
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
timm>=0.9.0

# 基础工具
setuptools>=69.0.0
wheel>=0.42.0 